+NewRouterReplyContent, Finished Connection Init
Added NewRouterReplyContent. Sent as a response to a NewRouterContent
packet. Code for connection initialisation is mainly finished. Routers
can be initialised and directly connected to other routers. Clients can
be directly connected to a router. New neighbouring routers are added to
a pendingRouters array list until a response from them is received. When
it is received they are added to a connectedRouters HashMap. This acts
like a DNS table - the router name is the key and the IP the value.
Similarly, connected clients are added to a connectedClients HashMap.