+Server onRecToType and onRecReg, UpdatedClient
Added code to server to handle regular and toType packets. Updated
fields in RoutingTableEntry (changed from InetAddess to
InetSocketAddress) and added getters and setters. Updated Client to
check if this type is equal to dst type for messages toType. Updated
code to check if input dst is to a specific type.